Refactor passage card into new component
