+++ client/anoixo-client/src/results/css/PassageCard.css
@@ -1,0 +1,8 @@
+.passage-card {
+  margin: 1.2rem;
+  padding: 0.8rem;
+}
+
+.matched-word {
+  background-color: lightblue;
+}
