+++ client/anoixo-client/src/results/css/ResultsListing.css
@@ -3,12 +3,3 @@
   margin-left: 1.2rem;
   display: flex;
 }
-
-.passage-card {
-  margin: 1.2rem;
-  padding: 0.8rem;
-}
-
-.matched-word {
-  background-color: lightblue;
-}
